requirements.txt
boto3>=1.34.0
Pillow>=10.0.0
requests>=2.28.0
